added support for exponents and added and error check for missing operands and invalid operator sequence

+++ src/calculations.cpp
@@ -35,7 +35,7 @@ namespace calculate
         }
     }
 
-    double exponent(int x, int y)
+    double exponent(double x, double y)
     {
         return pow(x, y);
     }
